+++ nodes/Klaviyo/TemplateFields.ts
@@ -1,0 +1,80 @@
+export const getTemplateAttrFields = [
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'name',
+		value: 'name'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'editor_type',
+		value: 'editor_type'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'html',
+		value: 'html'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'text',
+		value: 'text'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'amp',
+		value: 'amp'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'created',
+		value: 'created'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'updated',
+		value: 'updated'
+	},
+];
+
+export const templateSortValues = [
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'created (Oldest First)',
+		value: 'created'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'created (Newest First)',
+		value: '-created'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased, n8n-nodes-base/node-param-display-name-miscased-id
+		name: 'id (A-Z)',
+		value: 'id'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased, n8n-nodes-base/node-param-display-name-miscased-id
+		name: 'id (Z-A)',
+		value: '-id'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'name (A-Z)',
+		value: 'name'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'name (Z-A)',
+		value: '-name'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'updated (Oldest First)',
+		value: 'updated'
+	},
+	{
+		// eslint-disable-next-line n8n-nodes-base/node-param-display-name-miscased
+		name: 'updated (Newest First)',
+		value: '-updated'
+	},
+];